GOV-20707 <GET /api/v2/relationships> endpoint does not give any response when using the _simpleFilter parameter.

+++ d360.model/DataAccessLayer/repositories/RelationshipRepository.cs
@@ -996,7 +996,7 @@ namespace d360.model.DataAccessLayer
 			});
 
 			string simpleFiltersTempTablesQuery = "";
-			if (dbArgs.ParameterNames.Contains("simpleFilter"))
+			if (dbArgs.ParameterNames.Contains("simpleFilter") && simpleFilters.Count()>0)
 			{
 				simpleFilterTempTables.AppendLine("drop table if exists #TempSimpleFilter");
 				simpleFilterTempTables.AppendLine("create table #TempSimpleFilter(IntersectId int)");
